feat: use render functions instead of components
The using components it is difficult to pass through additional props
and causes remount issues when an arrow function is used instead of an
actual component with a type.

BREAKING CHANGE:
before:
You'd pass through destroyer props including the component, and a page
to render which can either be created or cloned.

after:
<PageGuard
  renderPage={() => (
    <MyComponent/>
  )}
  renderDestroyer={() => (
    <MyDestroyer/>
  )}
/>